Reshaped the form of MarketAnalyzers
I had originally envisioned passing all relevant info into the analyzer's
function call, but that just gets too hairy with some analysis (e.g. velocity
analysis). Instead, we're now passing basic before & after data in, which can
be useful for *most* analysis (but not all) and then if the analyzer needs extra
data it can pull it in itself based on the params.